Bug 834768: Fix signed/unsigned comparison build warnings in signaling_unittests.cpp. r=jesup
UltraBlame original commit: ec05a370f01d46722b1589d55223dd44d3eccb07

+++ media/webrtc/signaling/test/signaling_unittests.cpp
@@ -467,7 +467,7 @@ class ParsedSDP {
 
     
     std::string sdp;
-    for(int i=0; i < sdp_lines.size(); i++)
+    for (size_t i = 0; i < sdp_lines.size(); i++)
     {
       sdp += sdp_lines[i];
     }
@@ -760,7 +760,7 @@ void CreateAnswer(sipcc::MediaConstraints& constraints, std::string offer,
   void CloseReceiveStreams() {
     std::vector<nsDOMMediaStream *> streams =
                             pObserver->GetStreams();
-    for(int i=0; i < streams.size(); i++) {
+    for (size_t i = 0; i < streams.size(); i++) {
       streams[i]->GetStream()->AsSourceStream()->StopStream();
     }
   }
